feat: 实现企业级RBAC权限系统
- 修复Zustand持久化hydration竞态条件，解决401认证错误
- 移除基于重定向的导航，实现企业级路由守卫
- 新增Admin和CMDB默认页面，展示模块卡片
- 实现PermissionGuard组件进行路由级权限检查
- 添加角色互斥性验证，防止冲突角色分配
- 权限操作字段中文化显示
- 修复Transfer组件缺失rowKey导致的空列表问题
- 审计服务添加JWT认证和基于权限的日志过滤
- 认证服务路由添加权限中间件保护

+++ services/auth-service/internal/grpc/server.go
@@ -34,7 +34,7 @@ func (s *AuthServer) ValidateToken(ctx context.Context, req *pb.ValidateTokenReq
 
 	return &pb.ValidateTokenResponse{
 		Valid:    true,
-		UserId:   uint32(claims.UserID),
+		UserId:   uint64(claims.UserID),
 		Username: claims.Username,
 	}, nil
 }
@@ -47,7 +47,7 @@ func (s *AuthServer) GetUser(ctx context.Context, req *pb.GetUserRequest) (*pb.G
 	}
 
 	return &pb.GetUserResponse{
-		Id:       uint32(user.ID),
+		Id:       uint64(user.ID),
 		Username: user.Username,
 		Email:    user.Email,
 		FullName: user.FullName,
